KEA-140 Changed Announcement view and changed description to rich text
