+++ backend/config/sync/admin-role.admin-lrh30qnl.json
@@ -11,11 +11,12 @@
         "fields": [
           "Title",
           "Poster",
-          "Description",
           "HostingGroup",
           "Private",
           "Link.Label",
-          "Link.URL"
+          "Link.URL",
+          "Description",
+          "StartDate.Date"
         ],
         "locales": [
           "en"
@@ -42,11 +43,12 @@
         "fields": [
           "Title",
           "Poster",
-          "Description",
           "HostingGroup",
           "Private",
           "Link.Label",
-          "Link.URL"
+          "Link.URL",
+          "Description",
+          "StartDate.Date"
         ],
         "locales": [
           "en"
@@ -62,11 +64,12 @@
         "fields": [
           "Title",
           "Poster",
-          "Description",
           "HostingGroup",
           "Private",
           "Link.Label",
-          "Link.URL"
+          "Link.URL",
+          "Description",
+          "StartDate.Date"
         ],
         "locales": [
           "en"
