+++ backend/config/sync/admin-role.strapi-editor.json
@@ -11,11 +11,12 @@
         "fields": [
           "Title",
           "Poster",
-          "Description",
           "HostingGroup",
           "Private",
           "Link.Label",
-          "Link.URL"
+          "Link.URL",
+          "Description",
+          "StartDate.Date"
         ]
       },
       "conditions": []
@@ -28,11 +29,12 @@
         "fields": [
           "Title",
           "Poster",
-          "Description",
           "HostingGroup",
           "Private",
           "Link.Label",
-          "Link.URL"
+          "Link.URL",
+          "Description",
+          "StartDate.Date"
         ]
       },
       "conditions": []
@@ -45,11 +47,12 @@
         "fields": [
           "Title",
           "Poster",
-          "Description",
           "HostingGroup",
           "Private",
           "Link.Label",
-          "Link.URL"
+          "Link.URL",
+          "Description",
+          "StartDate.Date"
         ]
       },
       "conditions": []
