+++ backend/config/sync/core-store.plugin_content_manager_configuration_components##common.link.json
@@ -59,11 +59,11 @@
         [
           {
             "name": "Label",
-            "size": 6
+            "size": 4
           },
           {
             "name": "URL",
-            "size": 6
+            "size": 8
           }
         ]
       ]
