+++ backend/tests/helpers/constructor.js
@@ -1,6 +1,23 @@
 /**
  * @param {string | number} id
  */
+const constructAnnouncement = (id) => {
+    return {
+        data: {
+            "Title": "Test Announcement " + id,
+            "HostingGroup": null,
+            "Poster": null,
+            "Description": "Announcement Description " + id,
+            "Private": false,
+            "locale": "en",
+            "publishedAt": Date.now(),
+        }
+    };
+}
+
+/**
+ * @param {string | number} id
+ */
 const constructGroup = (id) => {
     return {
         data: {
@@ -109,5 +126,6 @@ module.exports = {
     constructUserInfo, 
     constructLeader,
     constructSupportTicket,
-    constructEvent
+    constructEvent,
+    constructAnnouncement
 }
